backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pgvector==0.3.6
pydantic==2.10.6
alembic==1.14.1
pypdf==5.2.0
httpx==0.28.1
feedparser==6.0.11
duckduckgo-search==6.3.7
python-dotenv==1.0.1
google-generativeai==0.8.5
ollama==0.4.7
pytesseract==0.3.13
PyMuPDF==1.26.4
Pillow==11.1.0
firebase-admin==7.1.0
